Reorganize plugins and widget folder for explorer plugin
