+++ spyder/plugins/explorer/widgets/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright © Spyder Project Contributors
+# Licensed under the terms of the MIT License
+# (see spyder/__init__.py for details)
+
+"""
+spyder.plugins.explorer.widgets
+===============================
+
+Widgets used for the Files and Directories Explorer Plugin.
+"""
